chore: move types to @modern-js/types
