+++ packages/builder/webpack-builder/package.json
@@ -76,6 +76,7 @@
     "@modern-js/babel-preset-app": "workspace:*",
     "@modern-js/inspector-webpack-plugin": "1.0.3",
     "@modern-js/server": "workspace:*",
+    "@modern-js/types": "workspace:*",
     "@modern-js/utils": "workspace:*",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.5",
     "@svgr/webpack": "6.2.1",
