+++ packages/server/server/src/index.ts
@@ -1,12 +1,8 @@
 import { DevServer as Server } from './server';
-import type {
-  DevServerOptions,
-  DevServerHttpsOptions,
-  ModernDevServerOptions,
-} from './types';
+import type { ModernDevServerOptions } from './types';
 
 export { Server };
-export type { DevServerOptions, DevServerHttpsOptions, ModernDevServerOptions };
+export type { ModernDevServerOptions };
 
 export default (options: ModernDevServerOptions): Promise<Server> => {
   if (options == null) {
