+++ packages/toolkit/types/server/index.d.ts
@@ -2,3 +2,4 @@ export * from './context';
 export * from './hook';
 export * from './utils';
 export * from './route';
+export * from './devServer';
